Add files via upload
Design a Database for a company to Manage all its projects.
1. Company has diverse offices in several countries, which manage and co-ordinate the project of that country.
Head office has a unique name, city, country, address, phone number and name of the director.
Every head office manages a set of projects with Project code, title, project starting and end date, assigned budget and name of the person in-charge. One project is formed by the set of operations that can affect to several cities.
We want to know what actions are realized in each city storing its name, country and number of inhabitants.



2. Design a database for a lending company which manages lending among people (p2p lending)
Lenders that lending money are registered with an Id, name and available amount of money for the financial operations. 
Borrowers are identified by their id and the company registers their name and a risk value depending on their personal situation.
When borrowers apply for a loan, a new loan code, the total amount, the refund deadline, the interest rate and its purpose are stored in database. 
Lenders choose the amount they want to invest in each loan. A lender can contribute with different partial amounts to several loans.


3. Design a database to maintain the menu of a restaurant.
Each course has its name, a short description, photo and final price.
Each course has categories characterized by their names, short description, name of the employee in-charge of them.
Besides the courses some recipes are stored. They are formed by the name of their ingredients, the required amount, units of measurements and the current amount in the store.